+++ AlchemistCode/Cards/Uncommon/Knitbone.cs
@@ -2,7 +2,6 @@ using Alchemist.AlchemistCode.Commands;
 using BaseLib.Utils;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
-using MegaCrit.Sts2.Core.Localization;
 using MegaCrit.Sts2.Core.ValueProps;
 
 namespace Alchemist.AlchemistCode.Cards.Uncommon;
@@ -18,12 +17,7 @@ public class Knitbone : AlchemistCard
         WithTips(_ => Mixing.MixTips());
     }
 
-    protected override void AddExtraArgsToDescription(LocString description)
-    {
-        base.AddExtraArgsToDescription(description);
-        description.Add("MixesPlayed",
-            IsMutable && CombatState != null ? $" ({Mixing.PlayedThisCombat(Owner)})" : "");
-    }
+    protected override bool ShowsMixesPlayed => true;
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {